+++ sdk/storage/azure-storage-blobs/src/blob_batch_client.cpp
@@ -9,10 +9,10 @@
 
 #include "azure/core/credentials.hpp"
 #include "azure/core/http/curl/curl.hpp"
+#include "azure/storage/blobs/version.hpp"
 #include "azure/storage/common/constants.hpp"
 #include "azure/storage/common/shared_key_policy.hpp"
 #include "azure/storage/common/storage_per_retry_policy.hpp"
-#include "azure/storage/common/storage_version.hpp"
 
 namespace Azure { namespace Storage { namespace Blobs {
 
@@ -91,7 +91,7 @@ namespace Azure { namespace Storage { namespace Blobs {
   {
     std::vector<std::unique_ptr<Azure::Core::Http::HttpPolicy>> policies;
     policies.emplace_back(std::make_unique<Azure::Core::Http::TelemetryPolicy>(
-        Details::c_BlobServicePackageName, BlobServiceVersion));
+        Details::c_BlobServicePackageName, Version::VersionString()));
     policies.emplace_back(std::make_unique<Azure::Core::Http::RequestIdPolicy>());
     for (const auto& p : options.PerOperationPolicies)
     {
@@ -132,7 +132,7 @@ namespace Azure { namespace Storage { namespace Blobs {
   {
     std::vector<std::unique_ptr<Azure::Core::Http::HttpPolicy>> policies;
     policies.emplace_back(std::make_unique<Azure::Core::Http::TelemetryPolicy>(
-        Details::c_BlobServicePackageName, BlobServiceVersion));
+        Details::c_BlobServicePackageName, Version::VersionString()));
     policies.emplace_back(std::make_unique<Azure::Core::Http::RequestIdPolicy>());
     for (const auto& p : options.PerOperationPolicies)
     {
@@ -174,7 +174,7 @@ namespace Azure { namespace Storage { namespace Blobs {
   {
     std::vector<std::unique_ptr<Azure::Core::Http::HttpPolicy>> policies;
     policies.emplace_back(std::make_unique<Azure::Core::Http::TelemetryPolicy>(
-        Details::c_BlobServicePackageName, BlobServiceVersion));
+        Details::c_BlobServicePackageName, Version::VersionString()));
     policies.emplace_back(std::make_unique<Azure::Core::Http::RequestIdPolicy>());
     for (const auto& p : options.PerOperationPolicies)
     {
@@ -239,8 +239,8 @@ namespace Azure { namespace Storage { namespace Blobs {
         requestBody += getBatchBoundary();
 
         auto blobUrl = m_serviceUrl;
-        blobUrl.AppendPath(subrequest.ContainerName);
-        blobUrl.AppendPath(subrequest.BlobName);
+        blobUrl.AppendPath(Details::UrlEncodePath(subrequest.ContainerName));
+        blobUrl.AppendPath(Details::UrlEncodePath(subrequest.BlobName));
         BlobRestClient::Blob::DeleteBlobOptions protocolLayerOptions;
         protocolLayerOptions.DeleteSnapshots = subrequest.Options.DeleteSnapshots;
         protocolLayerOptions.IfModifiedSince = subrequest.Options.AccessConditions.IfModifiedSince;
@@ -261,8 +261,8 @@ namespace Azure { namespace Storage { namespace Blobs {
         requestBody += getBatchBoundary();
 
         auto blobUrl = m_serviceUrl;
-        blobUrl.AppendPath(subrequest.ContainerName);
-        blobUrl.AppendPath(subrequest.BlobName);
+        blobUrl.AppendPath(Details::UrlEncodePath(subrequest.ContainerName));
+        blobUrl.AppendPath(Details::UrlEncodePath(subrequest.BlobName));
         BlobRestClient::Blob::SetBlobAccessTierOptions protocolLayerOptions;
         protocolLayerOptions.Tier = subrequest.Tier;
         protocolLayerOptions.RehydratePriority = subrequest.Options.RehydratePriority;