+++ sdk/storage/azure-storage-files-shares/src/share_directory_client.cpp
@@ -11,8 +11,8 @@
 #include "azure/storage/common/storage_common.hpp"
 #include "azure/storage/common/storage_per_retry_policy.hpp"
 #include "azure/storage/common/storage_retry_policy.hpp"
-#include "azure/storage/common/storage_version.hpp"
 #include "azure/storage/files/shares/share_file_client.hpp"
+#include "azure/storage/files/shares/version.hpp"
 
 namespace Azure { namespace Storage { namespace Files { namespace Shares {
 
@@ -24,8 +24,8 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
   {
     auto parsedConnectionString = Azure::Storage::Details::ParseConnectionString(connectionString);
     auto directoryUri = std::move(parsedConnectionString.FileServiceUri);
-    directoryUri.AppendPath(shareName);
-    directoryUri.AppendPath(directoryPath);
+    directoryUri.AppendPath(Storage::Details::UrlEncodePath(shareName));
+    directoryUri.AppendPath(Storage::Details::UrlEncodePath(directoryPath));
 
     if (parsedConnectionString.KeyCredential)
     {
@@ -46,7 +46,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
   {
     std::vector<std::unique_ptr<Azure::Core::Http::HttpPolicy>> policies;
     policies.emplace_back(std::make_unique<Azure::Core::Http::TelemetryPolicy>(
-        Azure::Storage::Details::c_FileServicePackageName, FileServiceVersion));
+        Azure::Storage::Details::c_FileServicePackageName, Version::VersionString()));
     policies.emplace_back(std::make_unique<Azure::Core::Http::RequestIdPolicy>());
     for (const auto& p : options.PerOperationPolicies)
     {
@@ -72,7 +72,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
   {
     std::vector<std::unique_ptr<Azure::Core::Http::HttpPolicy>> policies;
     policies.emplace_back(std::make_unique<Azure::Core::Http::TelemetryPolicy>(
-        Azure::Storage::Details::c_FileServicePackageName, FileServiceVersion));
+        Azure::Storage::Details::c_FileServicePackageName, Version::VersionString()));
     policies.emplace_back(std::make_unique<Azure::Core::Http::RequestIdPolicy>());
     for (const auto& p : options.PerOperationPolicies)
     {
@@ -98,7 +98,7 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
   {
     std::vector<std::unique_ptr<Azure::Core::Http::HttpPolicy>> policies;
     policies.emplace_back(std::make_unique<Azure::Core::Http::TelemetryPolicy>(
-        Azure::Storage::Details::c_FileServicePackageName, FileServiceVersion));
+        Azure::Storage::Details::c_FileServicePackageName, Version::VersionString()));
     policies.emplace_back(std::make_unique<Azure::Core::Http::RequestIdPolicy>());
     for (const auto& p : options.PerOperationPolicies)
     {
@@ -118,14 +118,14 @@ namespace Azure { namespace Storage { namespace Files { namespace Shares {
   DirectoryClient DirectoryClient::GetSubDirectoryClient(const std::string& subDirectoryName) const
   {
     auto builder = m_shareDirectoryUri;
-    builder.AppendPath(subDirectoryName);
+    builder.AppendPath(Storage::Details::UrlEncodePath(subDirectoryName));
     return DirectoryClient(builder, m_pipeline);
   }
 
   FileClient DirectoryClient::GetFileClient(const std::string& filePath) const
   {
     auto builder = m_shareDirectoryUri;
-    builder.AppendPath(filePath);
+    builder.AppendPath(Storage::Details::UrlEncodePath(filePath));
     return FileClient(builder, m_pipeline);
   }