+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Release History
 
+## 1.0.0-beta.5 (Unreleased)
+
+### Breaking Changes
+
+* Move header `azure/storage/blobs/blob.hpp` to `azure/storage/blobs.hpp`
+
+## 1.0.0-beta.4 (2020-10-16)
+
+### Bug Fixes
+
+* Unencoded Container/Blob name is now encoded.
+
 ## 1.0.0-beta.3 (2020-10-13)
 
 ### New Features